管理画面に投稿詳細画面を追加し、リンク先を修正
