+++ app/controllers/admin_panel/confirmations_controller.rb
