+++ app/controllers/admin_panel/dashboards_controller.rb
