+++ app/controllers/admin_panel/omniauth_callbacks_controller.rb
