+++ app/controllers/admin_panel/passwords_controller.rb
