+++ app/controllers/admin_panel/sessions_controller.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
-class Admin::SessionsController < Devise::SessionsController
+module AdminPanel
+  class SessionsController < Devise::SessionsController
 
   # before_action :configure_sign_in_params, only: [:create]
 
@@ -26,13 +27,14 @@ class Admin::SessionsController < Devise::SessionsController
   #   devise_parameter_sanitizer.permit(:sign_in, keys: [:attribute])
   # end
   
-  protected
-
-  def after_sign_in_path_for(resource)
-    admin_users_path
-  end
-
-  def after_sign_out_path_for(resource_or_scope)
-    new_admin_session_path
+    protected
+  
+    def after_sign_in_path_for(resource)
+      admin_users_path
+    end
+  
+    def after_sign_out_path_for(resource_or_scope)
+      new_admin_panel_session_path
+    end
   end
 end
