+++ app/controllers/admin_panel/users_controller.rb
@@ -1,6 +1,7 @@
-class Admin::UsersController < ApplicationController
-  layout 'admin'
-  
+module AdminPanel
+  class UsersController < ApplicationController
+    layout 'admin'
+    
     before_action :authenticate_admin!
     before_action :set_user, only: [:show, :destroy]
 
@@ -14,7 +15,7 @@ class Admin::UsersController < ApplicationController
     def destroy
         @user = User.find(params[:id])
         @user.destroy
-        redirect_to admin_users_path, notice: 'ユーザーを削除しました。'
+        redirect_to admin_panel_users_path, notice: 'ユーザーを削除しました。'
     end
     
     private
@@ -22,5 +23,5 @@ class Admin::UsersController < ApplicationController
     def set_user
       @user = User.find(params[:id])
     end
-    
+  end   
 end
