+++ app/views/admin_panel/comments/index.html.erb
@@ -17,10 +17,10 @@
           <tr>
             <td><%= comment.user.name %></td>
             <td><%= comment.content %></td>
-            <td><%= link_to comment.post.title, post_path(comment.post) %></td>
+            <td><%= link_to comment.post.title, admin_panel_post_path(comment.post) %></td>
             <td><%= comment.created_at.strftime("%Y-%m-%d %H:%M") %></td>
             <td class="text-center">
-              <%= link_to '削除', admin_comment_path(comment), method: :delete, data: { confirm: '本当に削除しますか？' }, class: 'btn btn-danger btn-sm' %>
+              <%= link_to '削除', admin_panel_comment_path(comment), method: :delete, data: { confirm: '本当に削除しますか？' }, class: 'btn btn-danger btn-sm' %>
             </td>
           </tr>
         <% end %>
