+++ app/views/admin_panel/confirmations/new.html.erb
