+++ app/views/admin_panel/mailer/confirmation_instructions.html.erb
