+++ app/views/admin_panel/mailer/unlock_instructions.html.erb
