+++ app/views/admin_panel/passwords/edit.html.erb
