+++ app/views/admin_panel/passwords/new.html.erb
