+++ app/views/admin_panel/registrations/edit.html.erb
