+++ app/views/admin_panel/registrations/new.html.erb
