+++ app/views/admin_panel/sessions/new.html.erb
