+++ app/views/admin_panel/shared/_links.html.erb
