+++ app/views/admin_panel/shared/_sidebar.html.erb
@@ -2,13 +2,13 @@
   <div class="sidebar-sticky">
     <ul class="nav flex-column mt-4">
       <li class="nav-item-side" style="margin: 10px 0;">
-        <%= link_to 'ユーザー一覧', admin_users_path, class: "nav-link" %>
+        <%= link_to 'ユーザー一覧', admin_panel_users_path, class: "nav-link" %>
       </li>
       <li class="nav-item-side" style="margin: 10px 0;">
-        <%= link_to 'カテゴリ管理', admin_categories_path, class: "nav-link" %>
+        <%= link_to 'カテゴリ管理', admin_panel_categories_path, class: "nav-link" %>
       </li>
       <li class="nav-item-side" style="margin: 10px 0;">
-        <%= link_to 'コメント一覧', admin_comments_path, class: "nav-link" %>
+        <%= link_to 'コメント一覧', admin_panel_comments_path, class: "nav-link" %>
       </li>
     </ul>
   </div>
