+++ app/views/admin_panel/users/index.html.erb
@@ -14,7 +14,7 @@
             <tr>
               <td><%= user.name %></td>
               <td><%= user.email %></td>
-              <td class="text-center"><%= link_to '詳細', admin_user_path(user), class: 'btn btn-primary btn-sm' %></td>
+              <td class="text-center"><%= link_to '詳細', admin_panel_user_path(user), class: 'btn btn-primary btn-sm' %></td>
             </tr>
           <% end %>
         </tbody>
