+++ app/views/layouts/admin.html.erb
@@ -12,13 +12,13 @@
   <body>
     <!--ヘッダー -->
     <header>
-      <%= render 'admin/shared/header' %>
+      <%= render 'admin_panel/shared/header' %>
     </header>
   
     <div class="container-fluid">
       <div class="row">
         <!-- サイドバー -->
-        <%= render 'admin/shared/sidebar' %>
+        <%= render 'admin_panel/shared/sidebar' %>
   
         <!-- メインコンテンツ -->
         <main role="main" class="col-md-8 mx-auto col-lg-8">
